Packet Interpreter server creating event

+++ server/inc/Engine/Event.hpp
@@ -5,18 +5,19 @@
 #ifndef RTYPE_EVENT_HPP
 #define RTYPE_EVENT_HPP
 
-namespace Engine {
-    struct Event
-    {
-        unsigned int _entityId;
-        float _xVelocity;
-        float _yVelocity;
-        bool _shotReleased;
-        bool _chargingShot;
-
-        Event() = delete;
-        explicit Event(unsigned int entityId);
-    };
+namespace Engine
+{
+        struct Event
+        {
+                unsigned int	_entityId;
+                short		_xVelocity;
+                short		_yVelocity;
+                bool		_shotReleased;
+                bool		_chargingShot;
+		
+                Event() = delete;
+                explicit Event(unsigned int entityId);
+        };
 }
 
 #endif //RTYPE_EVENT_HPP
